containers/app/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pyyaml==6.0.1
launchdarkly-server-sdk==9.12.0 